psos: core priority level must be signed integer

+++ include/psos/psos.h
@@ -365,8 +365,9 @@ u_long tm_wkwhen(u_long date,
 		 u_long time,
 		 u_long ticks);
 
-int psos_task_normalize_priority(unsigned long psos_prio);
-int psos_task_denormalize_priority(unsigned long core_prio);
+int psos_task_normalize_priority(u_long psos_prio);
+
+u_long psos_task_denormalize_priority(int core_prio);
 
 extern unsigned int psos_long_names;
